requirements.txt
streamlit
paho.mqtt